# Make the D Brainfuck example faster