@JackStouffer 2016-06-09
Also, you should see if you can't replace some of those arrays with `Appender!int`.