@wilzbach 2016-06-09
> Also, you should see if you can't replace some of those arrays with Appender!int.

@JackStouffer I profiled and it looks like this won't make a difference - 99.99% of the runtime are spent in `run()`

![image](https://cloud.githubusercontent.com/assets/4370550/15916454/8bc36c86-2df3-11e6-95d3-3c73ae8ed2ef.png)