@JackStouffer 2016-06-09
If the `to!char(int)` is taking your time, then just change it to `cast(char)`.