@kostya 2016-06-09
i imlement new bf version without hash_map, https://github.com/kostya/benchmarks/blob/master/brainfuck2/bf.cr
can you translate it to d?