@kostya 2016-06-17
now there is brainfuck2 benchmark https://github.com/kostya/benchmarks#brainfuck-v20, which not used hash_map at all.